requirements.txt
astroid==1.6.1
autopep8==1.3.4
blinker==1.4
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.2.1
Django==1.11.9
django-oauth-toolkit==1.0.0
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.7.7
gunicorn==19.7.1
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
isort==4.2.15
itypes==1.1.0
jedi==0.11.1
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mongoengine==0.15.0
oauthlib==2.0.6
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycodestyle==2.3.1
Pygments==2.2.0
pylint==1.8.2
pymongo==3.6.0
pytz==2017.3
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
wrapt==1.10.11
